warning 제거

+++ SmGraph/Experiment.hpp
@@ -4,13 +4,13 @@ namespace ex
 	void printState(Graph& graph)
 	{
 		int ecount = 0;
-		printf("node count:%d\n", graph.getNodeCount());
-		int s = graph.getNodeCount();
+		printf("node count:%d\n", (int)graph.getNodeCount());
+		int s = (int)graph.getNodeCount();
 		for (int i = 0; i < s; ++i)
-			ecount += graph.get(i)->getEdges().size();
+			ecount += (int)graph.get(i)->getEdges().size();
 
 		printf("edge count:%d\n", ecount);
-		printf("avg degree:%f\n", (double)ecount / graph.getNodeCount());
+		printf("avg degree:%f\n", (double)ecount / (int)graph.getNodeCount());
 	}
 	void printCost(Graph& origin, Graph& summarized)
 	{
